Adding back DUMB and reverting GME changes
Urgh..

Module music still doesn't work yet though.

+++ game-music-emu/gme/Fir_Resampler.cpp
@@ -1,4 +1,4 @@
-// Game_Music_Emu https://bitbucket.org/mpyne/game-music-emu/
+// Game_Music_Emu 0.6.0. http://www.slack.net/~ant/
 
 #include "Fir_Resampler.h"
 
@@ -186,8 +186,8 @@ int Fir_Resampler_::avail_( blargg_long input_count ) const
 
 int Fir_Resampler_::skip_input( long count )
 {
-	int remain = write_pos - buf.begin();
-	int max_count = remain - width_ * stereo;
+	int remain = int(write_pos - buf.begin());
+	int max_count = int(remain - width_ * stereo);
 	if ( count > max_count )
 		count = max_count;
